Django_02_SQL/requirement.txt
django==5.1.5
psycopg2-binary==2.9.10